+++ src/main/java/lotto/data/Database.java
@@ -1,7 +1,5 @@
 package lotto.data;
 
-import lotto.Lotto;
-
 import java.util.ArrayList;
 import java.util.List;
