+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import lotto.constants.ErrorCode;
+import lotto.data.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
